+++ src/test/java/org/opensearch/neuralsearch/processor/NeuralQueryEnricherProcessorIT.java
@@ -9,6 +9,7 @@ import static org.opensearch.neuralsearch.TestUtils.TEST_SPACE_TYPE;
 import static org.opensearch.neuralsearch.TestUtils.createRandomVector;
 
 import java.util.Collections;
+import java.util.List;
 import java.util.Map;
 
 import org.apache.hc.core5.http.io.entity.EntityUtils;
@@ -22,6 +23,7 @@ import org.opensearch.core.rest.RestStatus;
 import org.opensearch.neuralsearch.BaseNeuralSearchIT;
 import org.opensearch.neuralsearch.query.HybridQueryBuilder;
 import org.opensearch.neuralsearch.query.NeuralQueryBuilder;
+import org.opensearch.neuralsearch.query.NeuralSparseQueryBuilder;
 
 import com.google.common.primitives.Floats;
 
@@ -30,9 +32,11 @@ import lombok.SneakyThrows;
 public class NeuralQueryEnricherProcessorIT extends BaseNeuralSearchIT {
 
     private static final String index = "my-nlp-index";
+    private static final String sparseIndex = "my-nlp-index-sparse";
     private static final String search_pipeline = "search-pipeline";
     private static final String ingest_pipeline = "nlp-pipeline";
     private static final String TEST_KNN_VECTOR_FIELD_NAME_1 = "test-knn-vector-1";
+    private static final String TEST_RANK_FEATURES_FIELD_NAME_1 = "test-rank-features-1";
     private final float[] testVector = createRandomVector(TEST_DIMENSION);
 
     @Before
@@ -58,6 +62,25 @@ public class NeuralQueryEnricherProcessorIT extends BaseNeuralSearchIT {
             assertFalse(response.isEmpty());
         } finally {
             wipeOfTestResources(index, ingest_pipeline, modelId, search_pipeline);
+        }
+    }
+
+    @SneakyThrows
+    public void testNeuralQueryEnricherProcessor_whenNoModelIdPassedInNeuralSparseQuery_thenSuccess() {
+        String modelId = null;
+        try {
+            initializeIndexIfNotExist(sparseIndex);
+            modelId = prepareSparseEncodingModel();
+            createSearchRequestProcessor(modelId, search_pipeline);
+            createPipelineProcessor(modelId, ingest_pipeline, ProcessorType.SPARSE_ENCODING);
+            updateIndexSettings(sparseIndex, Settings.builder().put("index.search.default_pipeline", search_pipeline));
+            NeuralSparseQueryBuilder neuralSparseQueryBuilder = new NeuralSparseQueryBuilder();
+            neuralSparseQueryBuilder.fieldName(TEST_RANK_FEATURES_FIELD_NAME_1);
+            neuralSparseQueryBuilder.queryText("hello");
+            Map<String, Object> response = search(sparseIndex, neuralSparseQueryBuilder, 2);
+            assertFalse(response.isEmpty());
+        } finally {
+            wipeOfTestResources(sparseIndex, ingest_pipeline, modelId, search_pipeline);
         }
     }
 
@@ -118,6 +141,12 @@ public class NeuralQueryEnricherProcessorIT extends BaseNeuralSearchIT {
                 Collections.singletonList(TEST_KNN_VECTOR_FIELD_NAME_1),
                 Collections.singletonList(Floats.asList(testVector).toArray())
             );
+            assertEquals(1, getDocCount(indexName));
+        }
+
+        if (sparseIndex.equals(indexName) && !indexExists(indexName)) {
+            prepareSparseEncodingIndex(indexName, List.of(TEST_RANK_FEATURES_FIELD_NAME_1));
+            addSparseEncodingDoc(indexName, "1", List.of(TEST_RANK_FEATURES_FIELD_NAME_1), List.of(Map.of("hi", 1.0f, "hello", 1.1f)));
             assertEquals(1, getDocCount(indexName));
         }
     }